+++ core/types/hashing.go
@@ -30,8 +30,8 @@ import (
 	"github.com/erigontech/erigon-lib/crypto"
 	"github.com/erigontech/erigon-lib/crypto/cryptopool"
 	"github.com/erigontech/erigon-lib/rlp"
+	"github.com/erigontech/erigon-lib/rlphacks"
 	"github.com/erigontech/erigon-lib/trie"
-	"github.com/erigontech/erigon/turbo/rlphacks"
 )
 
 type DerivableList interface {
