+++ erigon-lib/rlphacks/struct.go
